MyChannel/Backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
google-cloud-aiplatform==1.66.0
google-cloud-pubsub==2.21.1
google-cloud-logging==3.11.3
python-dotenv==1.0.1


